Update schema: positions to float, add unit enums
Change GameObjectSchema.x/y from uint16 to float32 for precise positioning. Introduce generated UnitArchetype and UnitBehaviorState structs and add archetype/behaviorState, targetX/targetY fields to UnitSchema (reordering Colyseus type indexes accordingly). Update health/maxHealth/name type indexes and add new UnitType value (Sheep). These changes align the client schemas with updated server model and serialization indices.

+++ client/dino-islander/Assets/Scripts/Infrastructure/Schema/GameObjectSchema.cs
@@ -19,11 +19,11 @@ public GameObjectSchema() { }
 		[Type(0, "uint8")]
 		public byte type = default(byte);
 
-		[Type(1, "uint16")]
-		public ushort x = default(ushort);
+		[Type(1, "float32")]
+		public float x = default(float);
 
-		[Type(2, "uint16")]
-		public ushort y = default(ushort);
+		[Type(2, "float32")]
+		public float y = default(float);
 
 		[Type(3, "string")]
 		public string id = default(string);
